refactor: pick out get user infos logic to custom hook
